+++ src/main/java/lotto/exceptions/BadTicketException.java
@@ -1,4 +1,4 @@
-package lotto.domain.exceptions;
+package lotto.exceptions;
 
 public class BadTicketException extends RuntimeException {
