+++ src/main/java/lotto/exceptions/BadWinnerBallsException.java
@@ -1,4 +1,4 @@
-package lotto.domain.exceptions;
+package lotto.exceptions;
 
 public class BadWinnerBallsException extends RuntimeException {
