+++ src/main/java/lotto/exceptions/IllegalTypeEnteredException.java
@@ -1,4 +1,4 @@
-package lotto.view;
+package lotto.exceptions;
 
 public class IllegalTypeEnteredException extends RuntimeException{
