+++ src/test/java/lotto/domain/WinnerBallsTest.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.exceptions.BadWinnerBallsException;
+import lotto.exceptions.BadWinnerBallsException;
 import lotto.domain.strategies.ManuallyGeneratingStrategy;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
